+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 
 ## [1.6.0] - DEV
+
+### Added
+- Configurable compression timeout.
+
 ### Changed
 - Better workflow for headerbar.
 - Move saving state in the subtitle.